refactor: extract AppProviders from shared/index entry
Pulls the React provider chain (StrictMode + PostHog + Toast) out of
the UnifiedApp render block in `shared/index.tsx` and into a reusable
`<AppProviders>` component, so future iframe entry points (e.g. the
subtitle overlay) can mount their own React tree under the same
provider stack without re-running the entry-file side effects
(createRoot, style loaders, PostHog init).

To make AppProviders safely importable from unit tests, the
PostHogContext + PostHogProvider + usePostHog are moved to a new
side-effect-free module `src/contexts/PostHogContext.tsx`.
`shared/index.tsx` re-exports `PostHogProvider` and `usePostHog` from
that module to preserve backward compatibility with existing call
sites (e.g. `src/lib/analytics.ts`). The React context object remains
module-scoped, so all consumers continue to share a single context
instance.

No behaviour change.
